Bug with image update fixed and User/Section Controllers checked for exception handling and authorization

+++ Poodle/Poodle/Controllers/ProfileController.cs
@@ -5,6 +5,7 @@ using Poodle.Data.EntityModels;
 using Poodle.Services.Constants;
 using Poodle.Services.Contracts;
 using Poodle.Services.Dtos;
+using Poodle.Services.Exceptions;
 using Poodle.Services.Mappers;
 using System;
 using System.IO;
@@ -14,70 +15,79 @@ namespace Poodle.Web.Controllers
 {
     public class ProfileController : Controller
     {
-        private readonly IUsersService usersService;   
+        private readonly IUsersService usersService;
         private readonly UserMapper userMapper;
         private readonly IWebHostEnvironment webHostEnvironment;
 
         public ProfileController(IUsersService usersService, IWebHostEnvironment webHostEnvironment, UserMapper userMapper)
         {
-            this.usersService = usersService;          
+            this.usersService = usersService;
             this.webHostEnvironment = webHostEnvironment;
             this.userMapper = userMapper;
         }
-                
+
 
         public async Task<IActionResult> Index()
-        {            
+        {
             UserUpdateDto userToBeUpdated = this.userMapper.ConvertToUpdateDto(await this.GetLoggedUser());
             return this.View(userToBeUpdated);
         }
 
         [HttpPost]
-        public async Task <IActionResult> Edit(UserUpdateDto userModel)
+        public async Task<IActionResult> Edit(UserUpdateDto userModel)
         {
             if (!this.ModelState.IsValid)
             {
                 return this.View(userModel);
             }
 
-            var loggedUser = await this.GetLoggedUser();
-
-            //if any image is uploaded
-            if (userModel.ImageFile != null)
-            {                
-                //check if user has a pic other than default in the folder and delete it to add the new one
-                string currentUrl = loggedUser.Image.ImageUrl.Substring(1);
-                string path = Path.Combine(webHostEnvironment.WebRootPath, currentUrl);
-                string defaultPath = Path.Combine(webHostEnvironment.WebRootPath, ConstantsContainer.DEFAULT_IMAGEURL);
-
-                if (System.IO.File.Exists(path) & path != defaultPath)
-                {
-                    System.IO.File.Delete(path);
-                }
-                //upload new image
-
-                string folder = "img";
-                folder += Guid.NewGuid().ToString() + "_" + userModel.ImageFile.FileName;
-                string imageUrl = "/" + folder;
-                userModel.ImageUrl = imageUrl;
-                string serverFolder = Path.Combine(webHostEnvironment.WebRootPath, folder);
-                await userModel.ImageFile.CopyToAsync(new FileStream(serverFolder, FileMode.Create));
-
-                var updatedUser = await this.usersService.UpdateWeb(loggedUser.Id, userModel);
-
-                //update user in Context
-                this.ResetContext(updatedUser);
-                this.ViewData["SuccessMessage"] = "You updated your profile successfully";
-                return this.View(viewName: "Success");
-            }
-            else
+            try
             {
-                var updatedUser = await this.usersService.UpdateWeb(loggedUser.Id, userModel);
-                //update user in Context
-                this.ResetContext(updatedUser);
-                this.ViewData["SuccessMessage"] = "You updated your profile successfully";
-                return this.View(viewName: "Success");               
+                var loggedUser = await this.GetLoggedUser();
+
+                //if any image is uploaded
+                if (userModel.ImageFile != null)
+                {
+                    //check if user has a pic in the folder and delete it to add the new one
+                    string currentUrl = loggedUser.Image.ImageUrl.Substring(1);
+                    string path = Path.Combine(webHostEnvironment.WebRootPath, currentUrl);
+                    string defaultPath = Path.Combine(webHostEnvironment.WebRootPath, ConstantsContainer.DEFAULT_IMAGEURL.Substring(1));
+
+                    if (System.IO.File.Exists(path) && path != defaultPath)
+                    {
+                        System.IO.File.Delete(path);
+                    }
+
+                    //upload new image
+                    string folder = "img";
+                    folder += "/" + Guid.NewGuid().ToString() + "_" + userModel.ImageFile.FileName;
+                    string imageUrl = "/" + folder;
+                    userModel.ImageUrl = imageUrl;
+                    string serverFolder = Path.Combine(webHostEnvironment.WebRootPath, folder);
+                    await userModel.ImageFile.CopyToAsync(new FileStream(serverFolder, FileMode.Create));
+
+                    var updatedUser = await this.usersService.UpdateWeb(loggedUser.Id, userModel);
+
+                    //update user in Context
+                    this.ResetContext(updatedUser);
+                    this.ViewData["SuccessMessage"] = "You updated your profile successfully";
+                    return this.View(viewName: "Success");
+                }
+                else
+                {
+                    var updatedUser = await this.usersService.UpdateWeb(loggedUser.Id, userModel);
+                    //update user in Context
+                    this.ResetContext(updatedUser);
+                    this.ViewData["SuccessMessage"] = "You updated your profile successfully";
+                    return this.View(viewName: "Success");
+                }
             }
+            catch (DuplicateEntityException)
+            {
+                this.ModelState.AddModelError("Email", ConstantsContainer.USER_EXISTS);
+                return this.RedirectToAction(actionName: "Index", controllerName: "Profile");
+            }
+
 
         }
 
@@ -92,8 +102,13 @@ namespace Poodle.Web.Controllers
         {
             var userEmail = this.HttpContext.Session.GetString("CurrentUserEmail");
             var user = await this.usersService.GetByEmail(userEmail);
+
+            if (user == null)
+            {
+                this.RedirectToAction("Login", "Auth");
+            }
             return user;
         }
-                
+
     }
 }
